+++ gulpfile.js
@@ -275,7 +275,7 @@ const bundleModule = async () => {
 }
 
 const copyFiles = async() => {
-    const statics = [`lang`, `fonts`, `assets`, `icons`, `templates`, `packs`, `module.json`, `system.json`, `template.json`];
+    const statics = [`lang`, `languages`, `fonts`, `assets`, `icons`, `templates`, `packs`, `module.json`, `system.json`, `template.json`];
 
     const recursiveFileSearch = (dir, callback) => {
         const err = callback.err;
@@ -395,7 +395,7 @@ const buildWatch = () => {
     gulp.watch(`src/**/*.ts`, { ignoreInitial: false }, gulp.series(buildTS));
     gulp.watch(`src/**/*.less`, { ignoreInitial: false }, buildLess);
     gulp.watch(`src/**/*.sass`, { ignoreInitial: false }, buildSASS);
-    gulp.watch([`src/fonts`, `src/lang`, `src/templates`, `src/*.json`], { ignoreInitial: false }, copyFiles);
+    gulp.watch([`src/fonts`, `src/lang`, `src/languages`, `src/templates`, `src/*.json`], { ignoreInitial: false }, copyFiles);
 }
 
 /********************/
@@ -418,6 +418,7 @@ const clean = async () => {
     // if (fs.existsSync(path.join(`src`, mainFilePath))) { // MOD 4535992
         files.push(
             `lang`,
+            `languages`,
             `fonts`,
             `icons`,
             `packs`,